Fixed Naive Parsing of Content-Type HTTP header
It turns out checking for exactly "application/json" is a bad idea since
some servers/web frameworks include stuff like encoding in the header.
For example: Express.js since 4.16.x now set's the content-type header
to: 'application/json; charset=utf-8'.

To remedy this, we now check for the inclusion of application/json in
the string, but don't attempt to match the exact value